CB-38 Register a customer
Added a form for registration of new customers. The form needs som refactoring to increase the user experience. In its current form (no pun intended) there's way to many fields to fill out in one page. I'd like to slice it up in a more manageable way, so that the customer may do the registration in steps and not get a headache.

With that said, the form and its API-calls works just fine. A customer is indeed created in the DB if the input validation allows the user to proceed.

I also added a simple form for authenticating customers (right now the API-calls just include end-points for customers). This form and API-call works as well.