Created AuthContext, Refactor of NavBar-component
Added a AuthContext with the useContext-hook. The context stores data about the customerId, name and a boolean value for authentication-status.

The customerId is stored for later use in API-calls, while the name is stored for later use in MyPages for example.

Updated to NavBar-component to allow for login/logout action. Removed the "Signed in as...", and simply added a logout button if the user is authenticated. The name could be displayed in the MyPages later on.

The MyPages link in the navbar will now show the login prompt if the user is not authenticated, and if the user already is authenticated the user pages will render.