REFACTOR: Login-view, NavBar & routing
Made the login form component redirect the user back to the last visited page instead of always redirecting to the landing page. This way the user gets back to "Services" for example if the button on that page renders the login view, rather than sending the user to the landing page.

Made the link to "My pages" only visible to authenticated users, and also moved it to the right side of the navbar. Turned it to a button instead of a link as well.

Made some small changes to the routing to make sure that all the pages that should only be accessible by a user that is authenticated only renders if the isAuthenticated boolean is set to true. This way a user that's not authenticated cannot access these pages by manually writing the correct URL in the web browser - since they only exist in that context.