REFACTOR: Register-form
Decided to split the register form in two. First the customer enters the contact details, and then the data for the account.

Might make it multi-step later on, but at least this is a small improvement for the cognitive load.

Decided to put each field on a separate row to further decrease cognitive load.

Added the field for social security number (for a lack of an better translation of "personnummer").

Created a new context for holding the form-data while mounting/unmounting the different form components.