REFACTOR: GDPRCustomerData & FormEditCustomerData
Redesign of these components. Added some styling to the headings and made sure the components have a similar structure.

Made some changes to the html-elements for better heading-structure.

The edit-button was changed to a pen-symbol and moved into the container (added a descriptive aria-label for clarity).

Changes the colors of the buttons to be consistent.