REFACTOR: GDPRCustomerData
Split the view into several components. The customer data is now rendered from a different component, which in turn renders each field from another component. This allows for easier styling.

Removed the "DTO" and typed the customer in the component instead.

Also made the privacy modal a separate component, and moved the fetch into that component.