REFACTOR: FAQ.tsx
Made some accessibility changes to the F.A.Q-page. The map now returns buttons, instead of h4-elements to clarify that it is clickable. Added an aria-label that makes sure the user knows how to read the answers.

Also added an arrow for visual users to show that it is expandable. The first question also expands by default now when the page loads.

Changed the heading to a h1-element as well.