CREATED: ClosedJobsTable (WIP)
Added a table for displaying jobs that either has the status of CLOSED or APPROVED.

The messages returned from the server is just a long string, which doesn't present it self well. I decided to hide the messages by default and let the user decide if they want to read all the messages with a similar solution as the one used on the F.A.Q. page.

I applied some styling in the CSS-file that affected parts of the first table as well, but the CleaningsPerType-table still needs to be designed and reworked to let the action of approving or disapproving a job to cause the update of the jobs-data.