REFACTOR: MyPages / CleaningsPerType (WIP)
Added the choice of leaving a custom message in the modal when approving or disapproving a job.

Combined the API-call for approving/disapproving a job to one method and moved it to CustomerApi.

Extracted the sorting of the table to a separate component and translated it (SortJobByStatus.tsx)

Translated parts of the CleaningsPerType-component to Swedish.

Added an alert-message to replace the modal if the API-call somehow fails.

TODO:
1. Still needs to implement a dependency between this table and the other shown on MyPages to synchronize the fetch. Right now the second table with the finished jobs doens't automatically refresh the data when the customer approves or disapproves a job.
2. There's still some parts of the table data that probably should be translated.