REFACTOR: HTML-formatting of emails for contact us form (WIP)
Updated the API-call to expect a html-snippet. Made the method just use the email-component to generate a html snippet right now to see if it works, but this has to be adapted to actually include the data that the user enters in the form.

The logic works, and an email is sent with HTML-formatting. The image isn't live though, so it doesn't get rendered. This probably needs to be included in the email as an attachment.