REFACTOR: Moved context-data (plus some bugfixes)
All authentication-related data has been moved to the sessionStorage instead of the previous context. This to prevent someone from loosing their data while reloading the window, and allow for the iframe from Klarna to access data such as a username.

CREATED: Also created an error modal component that can be re-used when the server don't respond with a 2xx-status.

BUGFIX: Error message while booking a service
Added an error modal when someone tries to book a service on the same date as a previous job of the same type. Had to make some changes to the API-call to allow for this.