BUGFIX: Routing
Had to tweak the logic for handling the state of authentication to allow all browser to render the protected routes.

The login method now forces the window to reload, which isn't ideal performance-wise but ensures that the sessionStorage value can be read by the App.tsx and the routes to be rendered equally on all browsers.