BUGFIX: Pagination logic (CB-464)
The filter-function that excludes the jobs with the status of CLOSED or APPROVED wasn't considered in the pagination-logic and sometimes caused too many pages of jobs to be created even though they should fit on the same page.

The original pagination logic was based on the original array of jobs that contained more jobs than the table rendered. Moved the filter logic out of the table body, and based the pagination logic on the new variable that only contains the jobs that are rendered in the table.

Ideally we should only fetch the jobs that are needed for the table, but that's not how the end-point is set up at this moment.